scs_final/requirements.txt
flask
openai
python-dotenv
flask-cors
supabase
